oi-observability-lab/solution/01-otel-instrumentation/src/requirements-jsonservice.txt
waitress==2.1.0
xmltodict==0.12.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Werkzeug==2.0.3
click==8.0.4
flask==2.0.3
itsdangerous==2.1.1
Flask-HTTPAuth==4.5.0
#
prometheus-client==0.13.1
#
opentelemetry-distro==0.28b1
opentelemetry-api==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-exporter-otlp-proto-grpc==1.9.1
